Sonar fixes - Math operands should be cast before assignment - Suppress - Make sure that command line arguments are used safely here - Suppress - Replace this use of System.out or System.err by a logger.
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1876737 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hpsf/ClipboardData.java
@@ -36,7 +36,7 @@ public class ClipboardData {
 
     public void read( LittleEndianByteArrayInputStream lei ) {
         int offset = lei.getReadIndex();
-        int size = lei.readInt();
+        long size = lei.readInt();
 
         if ( size < 4 ) {
             String msg =
